Only cache /elections, /persons, /parties
When users visit /elections, /persons or /parties, they will receive a
cached response from Fastly. When they visit any other path, they won't.
This might need a bit of tweaking and tuning in future, as we see traffic
hit these paths.

Non-caching in this way involves setting the TTL for the cached object for
these paths to 0s, so that the response is never valid in the Fastly caches,
thus forcing Fastly to request the content from the origin.